change return value to float

+++ conkfinal.py
@@ -7,6 +7,8 @@ import matplotlib as mp
 import pandas as pd
 import seaborn as sns
 
+
+##Requirement 1.2 Read-In local csv using pandas ###
 df = pd.read_csv('assets\KPdata.csv')
 
 data = df
@@ -14,7 +16,9 @@ data = df
 ##Delete 2020 because NO TOURNAMENT due to COVID-19##
 ###Delete the 4 rows where the 2020 Final Four teams woud have been###
 data = data[data.Year != 2020]  
-avg = data.KPvalue.mean()
+avg = float(data.KPvalue.mean())
 print(avg)
-maxKPvalue = data[['KPvalue']].max()
+maxKPvalue = float(data[['KPvalue']].max())
 print(maxKPvalue)
+
+
